+++ newsy/admin.py
@@ -113,13 +113,17 @@ class NewsItemAdmin(ModelAdmin):
     fieldsets = [
         (None, {
             'fields': ['title', 'short_title', 'page_title', 'description',
-                       'sites', 'published', 'tags'],
+                       'published', 'tags'],
             'classes': ('general',),
         }),
         (_('Basic Settings'), {
             'fields': ['template'],
             'classes': ('low',),
             'description': _('Note: This page reloads if you change the selection. Save it first.'),
+        }),
+        (_('Sites'), {
+            'fields': ['sites'],
+            'classes': ('collapse',),
         }),
         (_('Advanced Settings'), {
             'fields': ['slug', 'publication_date'],