+++ newsy/forms.py
@@ -1,6 +1,7 @@
 from re import compile
 
 from django import forms
+from django.contrib.sites.models import Site
 from django.core.validators import RegexValidator
 from django.utils.translation import ugettext_lazy as _
 
@@ -20,10 +21,14 @@ class NewsItemAddForm(forms.ModelForm):
 class NewsItemForm(NewsItemAddForm):
     title = forms.CharField(widget=forms.TextInput(attrs={'size': 96}),
                             required=True)
-    tags = forms.CharField(widget=forms.TextInput(attrs={'size': 96}))
-    short_title = forms.CharField(widget=forms.TextInput(attrs={'size': 96}))
-    page_title = forms.CharField(widget=forms.TextInput(attrs={'size': 96}))
+    tags = forms.CharField(widget=forms.TextInput(attrs={'size': 96}),
+        required=False)
+    short_title = forms.CharField(widget=forms.TextInput(attrs={'size': 96}),
+        required=False)
+    page_title = forms.CharField(widget=forms.TextInput(attrs={'size': 96}),
+        required=False)
     slug = forms.CharField(required=True, validators=[validate_slug])
-    
+    sites = forms.ModelMultipleChoiceField(queryset=Site.objects.all(),
+        widget=forms.CheckboxSelectMultiple())
     class Meta(NewsItemAddForm.Meta):
         exclude = []